Move preferences Pinia store under stores/ui/ to mirror old Vuex layout
